requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.8
numpy==1.19.4
pandas==1.0.5
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.15.0
urllib3==1.25.11
dataiku-api-client==8.0.0
http://localhost:40000/public/packages/dataiku-internal-client.tar.gz

